Rename index and styles files
