+++ src/components/Shelf/ShelfPlaceholder/ShelfPlaceholder.less
@@ -1,0 +1,1 @@
+@import "../ShelfSlider/ShelfSlider";
