--- src/components/Shelf/ShelfPlaceholder/style.less
@@ -1,1 +1,0 @@
-@import "../ShelfSlider/style";
