+++ src/components/Shelf/ShelfSlider/ShelfSlider.less
