+++ src/editors/ShelfEditor/ShelfEditor.less
